Map DB models to r2h2; iNumCurrent integer
Add a mapping layer to populate r2h2 component instances from Django DB records and adjust related model and deps.

- Add dashboard migration to change ElectroCellPEM.iNumCurrent to PositiveIntegerField (default=1000) and update models.py accordingly.
- Introduce R2H2.map_to_db_objects() to load the first record from each Simulation M2M manager into the corresponding r2h2 component attribute and copy matching fields; call it at the start of run().
- Remove the direct from_django call in setUpElectro1 and keep build_curves() there.
- Minor view tweak to access the first electro cell's rR (likely to trigger lazy loading/inspection).
- Tighten pyproject Django constraint to "django>=5.2.12,<6" to avoid Django 6.x incompatibilities.

These changes enable r2h2 to be initialised from linked Django simulation objects and ensure the database field type and dependency versions are consistent.

+++ dashboard/models.py
@@ -100,7 +100,7 @@ class ElectroCellPEM(models.Model):
     arR_cell = models.JSONField(default=None, blank=True, null=True)
     arV_cell = models.JSONField(default=None, blank=True, null=True)
     # GEOMETRY & GRID
-    iNumCurrent = models.FloatField(default=1000)
+    iNumCurrent = models.PositiveIntegerField(default=1000)
     rA_cell = models.FloatField(default=1000.0, help_text="cm^2")
     rI_rated = models.FloatField(default=3.0, help_text="A/cm^2")
     # TEMPERATURE REFERENCES
